+++ rnalysis/general.py
@@ -371,3 +371,16 @@ def _get_attr_ref_path(ref):
         return read_attr_ref_table_path()
     else:
         return ref
+
+
+def _biotype_table_assertions(ref_df: pd.DataFrame):
+    """
+    Assert legality of Biotype Reference Table, and rename column names to standard names ('gene' and 'biotype').
+    :param ref_df:
+    :type ref_df:
+    :return:
+    :rtype:
+    """
+    assert ref_df.shape[
+               1] == 2, f"Invalid number of columns in Biotype Reference Table: found {ref_df.shape[1]} columns instead of 2!"
+    ref_df.rename(columns={ref_df.columns[0]: 'gene', ref_df.columns[1]: 'biotype'}, inplace=True)